highlight tool activation + prevent creating >1
1. Added activate option for highlight tool (as a side note, I could not
figure out how to activate the when the mouse is inside the rectangle).
Now, activation only occurs when hovering the handles.
2. I found it weird to be able to draw multiple highlights, but could
not see them unless I deleted the current one. So now, you can only draw
one rectangle.
3. The handles were drawn under the overlay, which made them difficult
to see/select. I moved the code that draws them last, so they will
appear on top of the overlay.