Throttle stack scrolls
Using underscore's throttle function, stack scrolls are throttled so that there is no more than one scroll every 50ms. This prevents too many load requests when the user scrolls the stack.

A better solution would be to throttle only when the images are not in the cache but I am not sure how to fit that into the existing framework.