feat(FreehandMouseTool and FreehandSculpterMouseTool): Updates the freehand suite to use spacing in
The FreehandMouseTool and FreehandMouseSculpter tool now use spacing based on image coordinates,
rather than the canvas. Thus the FreehandSculpter no longer adds a load of points if zoom right in
and sculpt. Additionally new points are only added to the region being sculpted. This improves
performance and reduces the number of superflous points.