perf(BrushTool, onNewImageBrushEventHandler, onImageRenderedBrushEventHandler): Only generate labelm
Now labelmaps are only added to the slice as they are created. If you try to erase a labelmap that
doesn't exist it now doesn't insert a blank one. This improves performance when stack scrolling
through a new set of images, as the memory for the labelmaps is now only allocated when you begin
drawing.

re #834