perf(BrushTool, drawBrush): Delete unused brush pixelData. Fix rendering flicker.
Now when you erase pixels from a pixelData mask, the pixelData memory is freed up if you erase the
last pixel of that mask on the slice. This is also makes it easier for tools like dcmjs to just
hoover up all the slices that have non-blank masks on them. Secondly I fixed a small flicker edge
case with the brush imageBitmapCache.