fix(BaseAnnotationTool.js): Passed an interaction type of 'mouse' while calling 'pointNearTool' from
Whenever the pointNearTool function was called inside the concrete ROI tool classes, the
interactionType was always 'undefined' for 'mousemove' events. But on the contrary, it was equal to
'mouse' when mousedown was fired. So, because of this the distance value was shifting between '15'
to '25' and causing the bug to appear.

fix #932