build(Build process and test running): Specifies a node version to use in package.json
There is currently no version of node specified in the project's package.json. Developers can rule
out issues related to using different node version if the project specifies which node version to
use. This change sets the version in package.json to that of lts/carbon (8.16.0), as well as adding
a .nvmrc file for those who use nvm to manage node versions. Using a node version older than
lts/carbon was causing drawHandles.test.js to fail in the tip of master.