v4.6.2-d: comments out function creating textbox content in RectangleROI
The RectangleROI tool sometimes throws reference errors for an element that we're trying to get the area from, for example in VideaCollect with older annotations. The annotation format had changed (as far as we can tell) such that Cornerstone does not seem fully backward compatible. So, since the textbox isn't drawn anyway, we disable creating the textbox coordinates altogether as well.