feat(imageIdSpecificStateManager): add `replace` imageIdToolState functions
There is currently no function to consume the same output from getImageIdToolState, which is the
state of a specific tool for the given imageId. addElementToolState requires a single item input,
and without replace, requires both a clear step and an iteration over all items to add into the
tool's id-specific state.