fix(CobbAngleTool, RectangleRoi): Fix the interaction of the tools with the TID.300 saving changes.
Correct notification on tool completion, with values restored correctly when loaded from TID.300 is
required for store/save functionality.  This is implemented in dcmjs, but the changes are required
here.  The cobb angle needed to be udpated after restoring, the Rectangle ROI needed the perimeter,
and changes needed to publish a completed measurement event.