fix(GlobalToolChangeHistory): implemented a Map datastructure to prevent duplication of history logs
Previously, for the GlobalToolChangeHistory state we had a simple array implementation, where for
each tool mode change a new object would be pushed to the history, the problem is that the
activation of tools like the LenghtTool can happen repeated times without any difference on their
parameters. This was causing some overriding of changes history, due to the fact that the history is
limited to 50 entries, so if you had some important change history to be saved and used a tool like
the Lenghtool 50 times you would lose the important  change you wanted to keep. Now this
implementation guarantees that only the most recent change will be saved on the history, prevent the
flood of the history from a single tool.